Added list of service providers on the invitation page.
When a admin invites a user to a team that user will now see
which service providers will have access to his data.
Also added a LetterOpener based on the ruby letter_opener gem that
will open HTML and TXT mails in a browser so you can see how the
email will look like.
Removed the profile groupzey-dev. This is now part of the dev profile.